Codechange: [OSX] Move OSX-specific font code to a separate file.
