+++ src/os/macosx/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_files(
     crashlog_osx.cpp
+    font_osx.cpp
     macos.h
     macos.mm
     osx_stdafx.h
